+++ CHANGELOG.md
@@ -18,6 +18,10 @@ All notable changes to this project will be documented in this file.
 - Remove `/broker/orders/match` endpoint, `matchOrders` must be done through the **Market API**. (#131)
 - Remove dead code in `IexecHubService` and `CommandStorage`. (#133)
 
+### Dependency Upgrades
+
+- Upgrade to `iexec-common` 8.4.0. (#135)
+
 ## [[8.3.0]](https://github.com/iExecBlockchainComputing/iexec-blockchain-adapter-api/releases/tag/v8.3.0) 2024-01-10
 
 ### New Features